+++ picopt/config/__init__.py
@@ -87,7 +87,9 @@ def _build_template() -> MappingTemplate:
                 {
                     "after": Optional(float),
                     "bigger": bool,
+                    "convert_jpeg_to_jxl": bool,
                     "convert_to": Optional(Sequence(Choice(convert_to_format_strs))),
+                    "convert_webp_to_jxl": bool,
                     "disable_programs": Sequence(str),
                     "dry_run": bool,
                     "extra_formats": Optional(Sequence(Choice(all_format_strs))),
@@ -484,7 +486,9 @@ def _settings_from_attrdict(ad: Any) -> PicoptSettings:
     return PicoptSettings(
         after=ad.after,
         bigger=ad.bigger,
+        convert_jpeg_to_jxl=ad.convert_jpeg_to_jxl,
         convert_to=tuple(ad.convert_to) if ad.convert_to is not None else None,
+        convert_webp_to_jxl=ad.convert_webp_to_jxl,
         disable_programs=tuple(ad.disable_programs),
         dry_run=ad.dry_run,
         extra_formats=tuple(ad.extra_formats) if ad.extra_formats is not None else None,